+++ cogwheels/helpers/settings.py
@@ -297,7 +297,7 @@ class BaseAppSettingsHelper:
         """
         if setting_name in self._deprecated_settings:
             depr = self._deprecated_settings[setting_name]
-            depr.warn_if_referenced_directly()
+            depr.warn_if_setting_attribute_referenced()
 
         if self.is_overridden(setting_name):
             return self.get_user_defined_value(setting_name)
@@ -305,10 +305,31 @@ class BaseAppSettingsHelper:
         if setting_name in self._replacement_settings:
             depr = self._replacement_settings[setting_name]
             if self.is_overridden(depr.setting_name):
-                depr.warn_if_deprecated_value_used_by_replacement()
+                depr.warn_if_user_using_old_setting_name()
                 return self.get_user_defined_value(depr.setting_name)
 
         return self.get_default_value(setting_name)
+
+    def is_value_from_deprecated_setting(self, setting_name):
+        """
+        Returns a boolean to help developers determine where a setting value
+        came from when dealing settings that replace deprecated settings.
+        Returns ``True`` when:
+
+        -   The setting named by ``setting_name`` is a replacement for a
+            deprecated setting.
+        -   The value returned by self.get_raw() for the setting comes from a
+            user-defined Django setting that uses the deprecated setting name
+        """
+        if not self.in_defaults(setting_name):
+            raise ValueError('%s is not a valid setting name' % setting_name)
+        if(
+            not self.is_overridden(setting_name) and
+            setting_name in self._replacement_settings
+        ):
+            depr = self._replacement_settings[setting_name]
+            return self.is_overridden(depr.setting_name)
+        return False
 
     def get_raw(self, setting_name, enforce_type=None):
         """