+++ cogwheels/tests/conf/settings.py
@@ -15,6 +15,12 @@ class TestAppSettingsHelper(BaseAppSettingsHelper):
             'REPLACED_SETTING_OLD',
             replaced_by='REPLACED_SETTING_NEW',
             warning_category=PendingDeprecationWarning,
+            additional_guidance=(
+                "The new setting offers much greater flexibility, whilst also "
+                "allowing developers to change X without changing Z. Check "
+                "out the version X.X release notes for further details: "
+                "https://your-django-project.readthedocs.io/en/latest/releases/X.X.html"
+            )
         ),
     )